refactor(events): share the auto-title template loader
`loadTitleTemplates` sat beside the Events quality hook, but the reminder
emails need the same templates to tell an auto-filled title from a
hand-written one. Move it into `@/lib/eventQuality` so the job doesn't
have to reach into a collection's internals.

Refs #611
