+++ src/lib/eventQuality/index.ts
@@ -17,18 +17,10 @@ export {
   MINIMUM_IMAGES,
   QUALITY_CHECK_VERSION,
 } from './checks'
-export {
-  EMAIL_RE,
-  findStaleDates,
-  GENERIC_TITLE_RE,
-  URL_RE,
-} from './heuristics'
-export {
-  buildEventQualityReport,
-  countOpenDocumentIssues,
-  type BuildReportOptions,
-} from './report'
+export { EMAIL_RE, findStaleDates, GENERIC_TITLE_RE, URL_RE } from './heuristics'
+export { buildEventQualityReport, countOpenDocumentIssues, type BuildReportOptions } from './report'
 export { SKIP_REASON_LABELS, shouldSkipQualityChecks } from './skip'
+export { loadTitleTemplates } from './titleTemplates'
 export type {
   CheckStatus,
   EventQualityInput,
